Continued implementation of exchange rate endpoint
* Changes in ExchangeRate:
  - changed `amountInRsd` type from Float to BigDecimal for better control over data
  - renamed `currency` to `code`
  - removed @NotNull annotation from `code` field since @NotBlank covers not null constraint
  - added @Builder, @AllArgsConstructor and @NoArgsConstructor lombok annotations

* Changes in ExchangeRateController:
  - added apiPrefix property to application.properties, and referenced it in @RequestMapping for this endpoint
  - injected ExchangeRateMapper through constructor (it was previously in ExchangeRateServiceLocalImpl)
  - implemented methods for CRUD operations (getAll, getById, create, update, delete)
  - added @Valid annotations to method arguments to check if received data is valid

* Changes in ExchangeRateMapper:
  - added @Component annotation so the class is auto-detected
  - added `id` in mapExchangeRatePutRequestToExchangeRate method signature

* Changes in ExchangeRatePostRequest:
  - added @AllArgsConstructor and @NoArgsConstructor lombok annotations
  - changed `amountInRsd` type from Float to BigDecimal
  - added validation for fields

* Changes in ExchangeRatePutRequest:
  - added @AllArgsConstructor and @NoArgsConstructor lombok annotations
  - changed `amountInRsd` type from Float to BigDecimal
  - added validation for fields

* Changes in ExchangeRateResponse:
  - changed `amountInRsd` type from Float to BigDecimal

+ Created ExchangeRateAlreadyDefined exception
+ Created ExchangeRateNotFound exception
+ Created ExchangeRateRepository
+ Created ExchangeRateService interface
+ Created ExchangeRateServiceLocalImpl:
  - implemented methods for CRUD operations (getAll, getById, create, update, delete)

+++ travel/src/main/java/com/tx/travel/payload/response/ExchangeRateResponse.java
@@ -1,16 +1,14 @@
 package com.tx.travel.payload.response;
 
-import lombok.AllArgsConstructor;
+import java.math.BigDecimal;
 import lombok.Builder;
 import lombok.Data;
-import lombok.NoArgsConstructor;
 
 @Data
 @Builder
 public class ExchangeRateResponse {
 
   private final Long id;
-//  private final String updatedAt;
   private final String code;
-  private final Float amountInRsd;
+  private final BigDecimal amountInRsd;
 }
